refactor(ten-vad): rename LstmState to ExtLstmState, update usage, and improve documentation consistency
