+++ crates/bunsen/src/blocks/rnn/lstm/ext_lstm_state.rs
